Function separated into own files for further re-organization.
main() in PSFModel.py
aperture handling in aperture.py
psf related functions in psf.py
